Project structure refactored.
