+++ .gitignore
@@ -4,4 +4,3 @@
 .project
 .checkstyle
 target
-bin
