--- core/.gitignore
@@ -1,6 +1,0 @@
-# Eclipse and Maven Files #
-.classpath
-.settings
-.project
-.checkstyle
-target
